workflow finished, adding tests
