+++ logging.yaml
@@ -1,0 +1,35 @@
+version: 1
+
+formatters:
+    simple:
+        format: "%(name)s - %(lineno)d -  %(message)s"
+
+    complex:
+        format: "%(asctime)s - %(name)s - %(lineno)d -  %(message)s"
+
+
+handlers:
+    console:
+        class: logging.StreamHandler
+        level: DEBUG
+        formatter: simple
+
+    file:
+        class: logging.handlers.TimedRotatingFileHandler
+        when: midnight
+        backupCount: 5
+        level: DEBUG
+        formatter: simple
+        filename : Thrift.log
+
+loggers:
+
+    qsoWidget:
+        level: INFO
+        handlers: [console,file]
+        propagate: yes
+
+    __main__:   
+        level: DEBUG
+        handlers: [console]
+        propagate: yes
