+++ ukpn/grafana/__init__.py
@@ -1,2 +1,2 @@
-"""Load Iter Datapipes"""
+"""Load Functions"""
 from .grafana_data_download import automate_csv_download
